Refactor content handling in ActivityPub models and components; introduce ContentValueResolver for improved content processing and add support for Editor.js and plain content types.
