+++ activity_pub/app/models/activity_pub/like.rb
@@ -1,6 +1,6 @@
 class ActivityPub::Like < ApplicationRecord
   belongs_to :source_ap_object, class_name: ActivityPub::Object.name
-  belongs_to :target_ap_object, class_name: ActivityPub::Object.name, counter_cache: true
+  belongs_to :target_ap_object, class_name: ActivityPub::Object.name, counter_cache: :likes_count
 
   scope :newest_first, ->{ order(created_at: :desc) }
 end
