+++ activity_pub/app/resources/activity_pub/article_resource.rb
@@ -10,7 +10,9 @@ module ActivityPub
     has_followers false
 
     register_data_attribute :name
+    register_data_attribute :content
     register_data_attribute :content_raw
+    register_data_attribute :content_mime_type
     register_data_attribute :url
     register_data_attribute :published_at
