+++ app/components/objects/content_article_component.html.erb
@@ -1,17 +1,20 @@
-<div class="mt-3">
-  <h1 class="font-medium text-xl mb-2"><%= object.name %></h1>
+<div class="flex justify-between border border-gray-100 shadow-sm h-44 overflow-hidden hover:bg-gray-50">
+  <div class="flex-1 h-44 overflow-hidden relative text-gray-600">
+    <div class="p-4 h-44">
+      <h1 class="font-medium text-xl mb-2"><%= object.name %></h1>
 
-  <div class="text-gray-600 text-sm leading-relaxed max-h-44 overflow-hidden relative [&_p]:mb-3">
-    <% if content.thumbnail %>
-      <div class="float-right w-32 h-32 overflow-hidden ml-4">
-        <%= image_tag content.thumbnail['url'] %>
+      <div class="text-sm leading-relaxed relative [&_p]:mb-3">
+        <%== content.excerpt.truncate(450) %>
       </div>
-    <% end %>
+    </div>
 
-    <%== content.excerpt.truncate(450) %>
-
-    <div class="pointer-events-none absolute bottom-0 left-0 right-0 h-12 bg-gradient-to-t from-white to-transparent"></div>
+    <div class="z-10 pointer-events-none absolute bottom-0 left-0 right-0 h-12 bg-gradient-to-t from-white to-transparent"></div>
   </div>
+
+  <% if content.thumbnail %>
+    <div class="w-40 overflow-hidden ml-4 h-full flex-shrink-0 bg-cover bg-center" style="background-image:url(<%= content.thumbnail['url'] %>)">
+    </div>
+  <% end %>
 </div>
 
 <%= render(interactions_component) %>
