+++ app/components/objects/content_article_component.rb
@@ -3,11 +3,7 @@
 module Objects
   class ContentArticleComponent < ContentBaseComponent
     def content
-      if @object.local?
-        EditorjsContentValue.new(@object.content_raw)
-      else
-        @object.content
-      end
+      Objects::ContentValueResolver.new(@object)
     end
   end
 end
