+++ app/value_objects/objects/plain_content_value.rb
@@ -1,0 +1,18 @@
+module Objects
+  class PlainContentValue < String
+    include ActionView::Helpers::SanitizeHelper
+
+    def sanitized
+      sanitize(self)
+    end
+
+    def excerpt
+      doc = Nokogiri::HTML.fragment(html)
+      doc.text
+    end
+
+    def thumbnail
+      nil
+    end
+  end
+end
