+++ app/views/panel/objects/_comment.html.erb
@@ -39,7 +39,8 @@
   </div>
 
   <div class="text-gray-600 text-sm leading-relaxed [&_p]:mb-3 [&_p:empty]:hidden">
-    <%= simple_format comment.data['content'] %>
+    <% content = Objects::ContentValueResolver.new(comment) %>
+    <%= content.sanitized %>
   </div>
 
   <%= render 'panel/comments/footer', comment: comment, liked_comment_ids: liked_comment_ids %>
